requirements.txt
pymongo
jupyter
ipykernel
python-dotenv == 0.20.0
pendulum
tqdm
pandas
matplotlib
seaborn
statsmodels
scipy
gcc=12.1.0
pandas-datareader
nltk
wordcloud
mplfinance
